+++ typescript/keyfunder/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @hyperlane-xyz/keyfunder
 
+## 0.1.43
+
+### Patch Changes
+
+- Updated dependencies [e7666ef]
+- Updated dependencies [5fac94d]
+- Updated dependencies [5899416]
+- Updated dependencies [5fac94d]
+  - @hyperlane-xyz/sdk@41.2.0
+  - @hyperlane-xyz/core@12.1.0
+  - @hyperlane-xyz/tron-sdk@24.1.3
+  - @hyperlane-xyz/metrics@0.2.36
+  - @hyperlane-xyz/utils@41.2.0
+
 ## 0.1.42
 
 ### Patch Changes